Adds logic to ensure that meal food deletion only removes the relation between meal and food.

+++ api/views.py
@@ -43,8 +43,12 @@ class FoodViews(viewsets.ViewSet):
 
     def destroy(self, request, food_id):
         food = get_object_or_404(Food, pk=food_id)
-        food.delete()
-        return HttpResponse(status=204)
+        import pdb; pdb.set_trace()
+        if len(food.meals.all()) > 0:
+            return HttpResponse(status=304)
+        else:
+            food.delete()
+            return HttpResponse(status=204)
 
 class MealViews(viewsets.ViewSet):
     def list(self, request):
@@ -69,8 +73,7 @@ class MealViews(viewsets.ViewSet):
     def destroy_meal_food(self, request, meal_id, id):
         meal = get_object_or_404(Meal, pk=meal_id)
         food = get_object_or_404(Food, pk=id)
-        meal_food = meal.foods.get(pk=id)
-        meal_food.delete()
+        meal.foods.remove(food)
         meal_name = str(meal.name)
         food_name = str(food.name)
         message = {"message": f"Successfully removed {food_name} from {meal_name}"}
